+++ ChangeLog.md
@@ -13,6 +13,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 - added ...
 
+### fixed
+
+- fixed execution time limit with sandbox
+
 
 
 ## [1.1.2] - 2018-12-10