+++ ChangeLog.md
@@ -7,12 +7,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 
 
-## [1.1.3] - 2018-12-xx
+## [1.1.3] - 2018-12-21
 
 ### added
 
 - added a cached thread pool to run the futures for scripts with execution 
         time limit
+- added function _time/in-range?_
 
 ### fixed